using partial now to improved UX
