+++ covalent/executor/base.py
@@ -40,43 +40,6 @@ app_log = logger.app_log
 log_stack_info = logger.log_stack_info
 
 
-def wrapper_fn(
-    function: TransportableObject,
-    call_before: List[Tuple[TransportableObject, TransportableObject, TransportableObject]],
-    call_after: List[Tuple[TransportableObject, TransportableObject, TransportableObject]],
-    *args,
-    **kwargs,
-):
-    """
-    Wrapper for serialized callable.
-
-    Execute preparatory shell commands before deserializing and
-    running the callable. This is the actual function to be sent to
-    the various executors.
-
-    """
-
-    for tup in call_before:
-        serialized_fn, serialized_args, serialized_kwargs = tup
-        cb_fn = serialized_fn.get_deserialized()
-        cb_args = serialized_args.get_deserialized()
-        cb_kwargs = serialized_kwargs.get_deserialized()
-        cb_fn(*cb_args, **cb_kwargs)
-
-    fn = function.get_deserialized()
-
-    output = fn(*args, **kwargs)
-
-    for tup in call_after:
-        serialized_fn, serialized_args, serialized_kwargs = tup
-        ca_fn = serialized_fn.get_deserialized()
-        ca_args = serialized_args.get_deserialized()
-        ca_kwargs = serialized_kwargs.get_deserialized()
-        ca_fn(*ca_args, **ca_kwargs)
-
-    return output
-
-
 class BaseExecutor(ABC):
     """
     Base executor class to be used for defining any executor
@@ -161,7 +124,7 @@ class BaseExecutor(ABC):
     @abstractmethod
     def execute(
         self,
-        function: TransportableObject,
+        function: Callable,
         args: List,
         kwargs: Dict,
         dispatch_id: str,
